Fix breaking change in HomeScreen, refactor styling
