+++ components/Product.js
@@ -1,11 +1,11 @@
 import React from 'react';
 import { Image, View } from 'react-native';
 
-import { ProductBody, ProductContainer } from '../styles';
+import { ProductBody } from '../styles/products';
 
 /**
  * @prop
- **/
+ * */
 
 // TODO @tommypoa to use styled-components
 function Product({ product }) {
