+++ components/StoreCard.js
@@ -1,11 +1,11 @@
-import React from "react";
-import { TouchableOpacity, View } from "react-native";
+import React from 'react';
+import { TouchableOpacity, View } from 'react-native';
 
-import { H3, Subtitle } from "../styles";
+import { H3, Subtitle } from '../styles/shared';
 
 /**
  * @prop
- **/
+ * */
 
 function StoreCard({ store, callBack }) {
   // Don't know if this is the best way or if there is a way to not pass in navigation as props
