+++ styles/shared.js
@@ -1,0 +1,31 @@
+import styled from 'styled-components/native';
+
+export const Title = styled.Text`
+  font-size: 24px;
+  font-weight: bold;
+  margin: 5% 5% 2.5%;
+`;
+export const Subtitle = styled.Text`
+  font-size: 16px;
+  font-weight: bold;
+  margin: 1% 5%;
+`;
+
+export const H3 = styled.Text`
+  font-size: 12px;
+  margin: 1% 6.5%;
+`;
+
+export const ScrollCategory = styled.Text`
+  font-size: 16px;
+  font-weight: bold;
+`;
+
+export const Button = styled.TouchableOpacity`
+  flex: 1 1 0;
+`;
+
+export const ButtonText = styled.Text`
+  color: white;
+  padding: 7.5px 7.5px;
+`;
