+++ styles/stores.js
@@ -1,0 +1,15 @@
+import styled from 'styled-components/native';
+
+export const StoreModalBar = styled.View`
+  border: 1px solid black;
+  background-color: black
+  height: 5px;
+  width: 50px;
+  border-radius: 5px;
+  margin: 2% 0% 0% 42%;
+`;
+
+export const StoreModal = styled.View`
+  background-color: #fff;
+  opacity: 0.8;
+`;
